+++ conftest.py
@@ -1,0 +1,4 @@
+
+def pytest_configure():
+    from bumpr import log
+    log.declare()